fix: The displayed circle that indicates anchor.maxRadius now has correct diameter
The displayed circle that indicates anchor.maxRadius now has correct diameter. Before setting a radius of e.g. 1852m (1nm) resulted in a displayed circle with a radius of about 1000m. It seems that aol-geometry-circle internally assumes that the radius is given in nautical miles, although the internal units of sk are SI-units (radius [m]). For the fix the argument is scaled by a factor of 1.852.